requirements.txt
Pillow==10.0.0
openpyxl==3.1.2
reportlab==4.0.4
tkcalendar==1.6.1